backend/requirements.txt
fastapi==0.115.0
hypercorn[h2]==0.17.3
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
httpx==0.27.2
structlog==24.4.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-multipart==0.0.12
argon2-cffi==23.1.0
email-validator==2.2.0
